pyproject.toml
[tool.poetry]
name = "salary_service"
version = "0.1.0"
description = "Salary Information Service"
authors = ["Your Name <you@example.com>"]
packages = [{include = "salary_service"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.2"
starlette = "0.36.3"
uvicorn = "^0.27.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
sqlalchemy = "^2.0.25"
pydantic-settings = "^2.10.1"
email-validator = "^2.2.0"
bcrypt = "^4.3.0"
python-multipart = "^0.0.20"
httpx = "^0.27.0"
pytest = "^8.4.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"